requirements.txt
Beaker==1.11.0
bottle==0.12.19
bottle-beaker==0.1.3
bottle-session==1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
freeze==3.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==2.13.5
peewee==2.10.1
PyJWT==1.5.2
redis==3.5.3
requests==2.25.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
